Applying real scenarios
Have to change the code to accept real scenarios, not the ones given by the user. For that most of the inputs code must be changed. Also creating scenarios. Currently is not working stacked when appliying functions to calculate uncertainties.  The error is in **Photodetector_func** in **inputs.lidar_inp.Lidar_inputs['Wavelength'][ind_UQ_PHOTO]** because it is not included in the scenarios and therefore doesn't have the appropriated length